add some comments | fix the directory 'screenshots' issue | fix the issue creatingAccount double information : add some clear field methods | delete not useful methods and comments

+++ src/test/java/testUtils/Screenshot.java
@@ -10,15 +10,16 @@ import org.apache.commons.io.FileUtils;
 class Screenshot {
 
     static void clearScreenshotDirectory(){
-        try {
-            FileUtils.cleanDirectory(new File("screenshots"));
-        } catch (IOException e) {
-            e.printStackTrace();
+        if(new File("screenshots").exists()) {
+            try {
+                FileUtils.cleanDirectory(new File("screenshots"));
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
         }
     }
 
     public static String getScreenshotName(String testname){
-        //String timestamp = String.valueOf(new Date().getTime());
         String timeFormat = new SimpleDateFormat("HH-mm-ss.SSS").format(new Date());
         return testname+"_"+timeFormat+".png";
     }
